Make some validator constructor public and use interfaces as parameters
