+++ validator/src/main/java/org/icann/rdapconformance/validator/configuration/ConfigurationFileParser.java
@@ -1,14 +1,9 @@
 package org.icann.rdapconformance.validator.configuration;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
 import java.io.IOException;
 import java.io.InputStream;
 
-public class ConfigurationFileParser {
+public interface ConfigurationFileParser {
 
-  private final ObjectMapper objectMapper = new ObjectMapper();
-
-  public ConfigurationFile parse(InputStream configuration) throws IOException {
-    return objectMapper.readValue(configuration, ConfigurationFile.class);
-  }
+  ConfigurationFile parse(InputStream configuration) throws IOException;
 }
