+++ validator/src/main/java/org/icann/rdapconformance/validator/workflow/profile/tig_section/general/TigValidation1Dot8.java
@@ -15,6 +15,7 @@ import org.icann.rdapconformance.validator.workflow.profile.tig_section.general.
 import org.icann.rdapconformance.validator.workflow.rdap.RDAPDatasetService;
 import org.icann.rdapconformance.validator.workflow.rdap.RDAPValidationResult;
 import org.icann.rdapconformance.validator.workflow.rdap.RDAPValidatorResults;
+import org.icann.rdapconformance.validator.workflow.rdap.RDAPValidatorResultsImpl;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.xbill.DNS.AAAARecord;
@@ -185,7 +186,7 @@ public final class TigValidation1Dot8 extends ProfileValidation {
       }
 
       String ipAddressJson = String.format("{\"ip\": \"%s\"}", ipAddress.getHostAddress());
-      SchemaValidator validator = new SchemaValidator(schema.path(), new RDAPValidatorResults(),
+      SchemaValidator validator = new SchemaValidator(schema.path(), new RDAPValidatorResultsImpl(),
           datasetService);
       return !validator.validate(ipAddressJson);
     }
