+++ validator/src/main/java/org/icann/rdapconformance/validator/workflow/rdap/RDAPValidatorResults.java
@@ -1,57 +1,25 @@
 package org.icann.rdapconformance.validator.workflow.rdap;
 
-import java.util.HashSet;
 import java.util.Set;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
-public class RDAPValidatorResults {
+public interface RDAPValidatorResults {
 
-  private static final Logger logger = LoggerFactory.getLogger(RDAPValidatorResults.class);
 
-  private final Set<RDAPValidationResult> results = new HashSet<>();
+  void add(RDAPValidationResult result);
 
-  private final Set<String> groups = new HashSet<>();
-  private final Set<String> groupErrorWarning = new HashSet<>();
+  Set<RDAPValidationResult> getAll();
 
-  public void add(RDAPValidationResult result) {
-    if (this.results.add(result)) {
-      logger.debug("adding error result {}", result);
-    }
-  }
+  boolean isEmpty();
 
-  public Set<RDAPValidationResult> getAll() {
-    return results;
-  }
+  Set<String> getGroupOk();
 
-  public boolean isEmpty() {
-    return results.isEmpty();
-  }
+  void addGroups(Set<String> groups);
 
-  public Set<String> getGroupOk() {
-    Set<String> groupsCopy = new HashSet<>(groups);
-    groupsCopy.removeAll(groupErrorWarning);
-    return groupsCopy;
-  }
+  Set<String> getGroupErrorWarning();
 
-  public void addGroups(Set<String> groups) {
-    this.groups.addAll(groups);
-  }
+  void addGroup(String group);
 
-  public Set<String> getGroupErrorWarning() {
-    return groupErrorWarning;
-  }
+  void addGroupErrorWarning(String group);
 
-  public void addGroup(String group) {
-    this.groups.add(group);
-  }
-
-  public void addGroupErrorWarning(String group) {
-    this.addGroup(group);
-    this.groupErrorWarning.add(group);
-  }
-
-  public Set<String> getGroups() {
-    return groups;
-  }
+  Set<String> getGroups();
 }
