+++ validator/src/test/java/org/icann/rdapconformance/validator/schemavalidator/RDAPDatasetServiceIt.java
@@ -3,6 +3,7 @@ package org.icann.rdapconformance.validator.schemavalidator;
 import org.icann.rdapconformance.validator.workflow.FileSystem;
 import org.icann.rdapconformance.validator.workflow.LocalFileSystem;
 import org.icann.rdapconformance.validator.workflow.rdap.RDAPDatasetService;
+import org.icann.rdapconformance.validator.workflow.rdap.RDAPDatasetServiceImpl;
 import org.icann.rdapconformance.validator.workflow.rdap.dataset.model.BootstrapDomainNameSpace;
 import org.icann.rdapconformance.validator.workflow.rdap.dataset.model.MediaTypes;
 import org.testng.annotations.BeforeMethod;
@@ -17,7 +18,7 @@ public class RDAPDatasetServiceIt  {
     @BeforeMethod
     public void setUp() {
         this.fileSystem = new LocalFileSystem();
-        this.rdapDatasetService = new RDAPDatasetService(this.fileSystem);
+        this.rdapDatasetService = new RDAPDatasetServiceImpl(this.fileSystem);
     }
 
     @Test
